Remove Symmetric Encryption Functionality
Symmetric encryption was originally supported by git-chat-message but
has been removed in this revision, since this functionality makes
git-chat-read implementation quite challenging.

Minor refactors and improvements.
