+++ include/gnupg/gpg-common.h
