+++ src/builtin/import-key.c
@@ -2,8 +2,8 @@
 #include <fcntl.h>
 
 #include "parse-options.h"
-#include "gpg-common.h"
-#include "git.h"
+#include "gnupg/gpg-common.h"
+#include "git/git.h"
 #include "working-tree.h"
 #include "utils.h"
 #include "fs-utils.h"
