+++ src/gnupg/encryption.c
@@ -2,9 +2,7 @@
 #include <string.h>
 #include <errno.h>
 
-#include "encryption.h"
-
-static gpgme_error_t gpgme_passphrase_cb(void *, const char *, const char *, int, int);
+#include "gnupg/encryption.h"
 
 void asymmetric_encrypt_plaintext_message(struct gc_gpgme_ctx *ctx,
 		const struct strbuf *message, struct strbuf *output,
@@ -47,7 +45,7 @@ void asymmetric_encrypt_plaintext_message(struct gc_gpgme_ctx *ctx,
 
 	// encrypt plaintext, always trusting gpg keys, and do not use default recipient
 	err = gpgme_op_encrypt(ctx->gpgme_ctx, keys_array, GPGME_ENCRYPT_ALWAYS_TRUST | GPGME_ENCRYPT_NO_ENCRYPT_TO,
-			message_in,  message_out);
+			message_in, message_out);
 	if (err) {
 		if (gpgme_err_code(err) == GPG_ERR_INV_VALUE)
 			BUG("invalid pointer passed to gpgme_op_encrypt(...)");
@@ -88,98 +86,4 @@ void asymmetric_encrypt_plaintext_message(struct gc_gpgme_ctx *ctx,
 
 	LOG_INFO("Successfully encrypted message");
 	errno = errsv;
-}
-
-void symmetric_encrypt_plaintext_message(struct gc_gpgme_ctx *ctx,
-		struct strbuf *message, struct strbuf *output, const char *passphrase)
-{
-	gpgme_error_t err;
-	int errsv = errno;
-
-	LOG_INFO("Encrypting plaintext message");
-
-	// if passphrase is defined, configure a passphrase callback to skip the
-	// default pinentry method
-	if (passphrase) {
-		err = gpgme_set_pinentry_mode(ctx->gpgme_ctx, GPGME_PINENTRY_MODE_LOOPBACK);
-		if (err)
-			GPG_FATAL("unable to set gpg pinentry mode to GPGME_PINENTRY_MODE_LOOPBACK", err);
-
-		gpgme_set_passphrase_cb(ctx->gpgme_ctx, gpgme_passphrase_cb, (char *)passphrase);
-	}
-
-	// build gpg data buffers for the plaintext input and ciphertext output
-	struct gpgme_data *message_in;
-	struct gpgme_data *message_out;
-	err = gpgme_data_new_from_mem(&message_in, message->buff, message->len, 0);
-	if (err)
-		GPG_FATAL("unable to create GPGME memory data buffer from plaintext message", err);
-
-	err = gpgme_data_new(&message_out);
-	if (err)
-		GPG_FATAL("unable to create GPGME data buffer for encrypted ciphertext", err);
-
-	err = gpgme_op_encrypt(ctx->gpgme_ctx, NULL, GPGME_ENCRYPT_NO_ENCRYPT_TO, message_in, message_out);
-	if (err) {
-		if (gpgme_err_code(err) == GPG_ERR_INV_VALUE)
-			BUG("invalid pointer passed to gpgme_op_encrypt(...)");
-
-		if (gpgme_err_code(err) == GPG_ERR_UNUSABLE_PUBKEY) {
-			struct _gpgme_op_encrypt_result *result = gpgme_op_encrypt_result(ctx->gpgme_ctx);
-
-			fprintf(stderr, "cannot encrypt message with invalid recipient:\n");
-
-			struct _gpgme_invalid_key *invalid_key = result->invalid_recipients;
-			while (invalid_key) {
-				fprintf(stderr, "\t%s - Reason: %s\n", invalid_key->fpr, gpgme_strerror(invalid_key->reason));
-
-				invalid_key = invalid_key->next;
-			}
-		}
-
-		GPG_FATAL("GPGME unable to encrypt message", err);
-	}
-
-	// read ciphertext into message_out strbuf
-	int ret = gpgme_data_seek(message_out, 0, SEEK_SET);
-	if (ret)
-		GPG_FATAL("failed to seek to beginning of gpgme data buffer", err);
-
-	char temporary_buffer[1024];
-	while ((ret = gpgme_data_read(message_out, temporary_buffer, 1024)) > 0)
-		strbuf_attach(output, temporary_buffer, ret);
-
-	if (ret < 0)
-		GPG_FATAL("failed to read from gpgme data buffer", err);
-
-	gpgme_data_release(message_in);
-	gpgme_data_release(message_out);
-
-	LOG_INFO("Successfully encrypted message");
-	errno = errsv;
-}
-
-/**
- * Passphrase entry callback to bypass default gpg pinentry method. Accepts a
- * passphrase via the `hook` argument, and write to the given file descriptor
- * `fd`, as described in the gpgme documentation.
- * */
-static gpgme_error_t gpgme_passphrase_cb(void *hook, const char *uid_hint,
-		const char *passphrase_info, int prev_was_bad, int fd)
-{
-	(void) uid_hint;
-	(void) passphrase_info;
-
-	const char *passphrase = (const char *) hook;
-
-	if (prev_was_bad)
-		return gpg_error(GPG_ERR_CANCELED);
-
-	ssize_t bytes_written = recoverable_write(fd, passphrase, strlen(passphrase));
-	bytes_written += recoverable_write(fd, "\n", 1);
-
-	if (bytes_written != (strlen(passphrase) + 1))
-		return gpg_error(GPG_ERR_CANCELED);
-
-	return 0;
 }
